refactor: create utils for helper text
