+++ src/index.tsx
@@ -38,7 +38,7 @@ export { default as Dialog } from './components/Dialog/Dialog';
 export { default as Divider } from './components/Divider';
 export { default as FAB } from './components/FAB';
 export { default as AnimatedFAB } from './components/FAB/AnimatedFAB';
-export { default as HelperText } from './components/HelperText';
+export { default as HelperText } from './components/HelperText/HelperText';
 export { default as IconButton } from './components/IconButton/IconButton';
 export { default as Menu } from './components/Menu/Menu';
 export { default as Modal } from './components/Modal';
@@ -108,7 +108,7 @@ export type { Props as DrawerItemProps } from './components/Drawer/DrawerItem';
 export type { Props as DrawerSectionProps } from './components/Drawer/DrawerSection';
 export type { Props as FABProps } from './components/FAB/FAB';
 export type { Props as FABGroupProps } from './components/FAB/FABGroup';
-export type { Props as HelperTextProps } from './components/HelperText';
+export type { Props as HelperTextProps } from './components/HelperText/HelperText';
 export type { Props as IconButtonProps } from './components/IconButton/IconButton';
 export type { Props as ListAccordionProps } from './components/List/ListAccordion';
 export type { Props as ListAccordionGroupProps } from './components/List/ListAccordionGroup';
